refactor(user.repository.ts): move s3 dependency into user service
